requirements.txt
# Retell SDK
retell-sdk>=1.0.0
aiosignal==1.3.2
aiosqlite==0.21.0
alembic==1.10.3
anyio==4.8.0
asyncpg==0.30.0
attrs==25.3.0
celery==5.3.6
certifi==2025.1.31
click==8.1.8
dnspython==2.7.0
ecdsa==0.19.1
email-validator==2.0.0
Faker==37.0.2
fastapi==0.95.0
frozenlist==1.5.0
greenlet==3.1.1
h11==0.14.0
httpcore==0.17.3
httpx==0.24.0
idna==3.10
iniconfig==2.0.0
Mako==1.3.9
MarkupSafe==3.0.2
multidict==6.1.0
neon==0.1.2
packaging==24.2
passlib==1.7.4
pluggy==1.5.0
propcache==0.3.0
psycopg2-binary==2.9.9
pyasn1==0.6.1
pydantic==1.10.7
pytest==8.3.5
pytest-asyncio==0.25.3
python-dotenv==1.0.0
python-jose==3.3.0
python-multipart==0.0.6
redis==5.0.1
rsa==4.9
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.9
starlette==0.26.1
typing_extensions==4.12.2
tzdata==2025.2
uvicorn==0.21.1
yarl==1.18.3
